# openapi.yaml
openapi: 3.0.1
info:
  title: personal_budget_v2
  version: 1.0.0
  description: >-
    Codecademy project building a personal budget application using node,
    express, postgresql, and js
paths:
...
  /categories/{id}:
    summary: Get single category envelope
    description: Work with user selected category, selected by ID parameter
    get:
      summary: GET category by ID
      description: Get single category envelope by user selected ID
      parameters:
        - name: id
          in: path
          description: ID of pet to return
          required: true
          schema:
            type: integer
            format: int64
      operationId: getCategoryById
      responses:
        '200':
          description: Default error sample response
          content:
            application/json:
              examples:
                return single category by id:
                  value: >-
                    {id: 1, name: "Petrol", assigned_budget: 200,
                    remaining_budget: 50}
    delete:
      summary: Delete category by ID
      description: >-
        User selected ID will delete matching category from table and cascade
        collapse matching expenses table
      parameters:
        - name: id
          in: path
          description: ID of pet to return
          required: true
          schema:
            type: integer
            format: int64
      operationId: deleteCategory
      responses:
        '204':
          description: Default error sample response
          content:
            application/json:
              examples:
                deleted:
                  value: '{deleted: "Deleted successfully!"}'
    put:
      summary: Update category
      description: >-
        Category is selected using ID as identifier. Category is then updated in
        the category table from supplued json input
      parameters:
        - name: id
          in: path
          description: ID of category to return
          required: true
          schema:
            type: integer
            format: int64
      operationId: updateCategory
      responses:
        '200':
          description: Default error sample response
          content:
            application/json:
              examples:
                updated successfully:
                  value: '{Updated: "Updated successfully!"}'
...
  /expenses/{id}:
    summary: expensed by id
    description: Work with individual expenses selected by ID
    parameters:
      - name: id
        in: path
        description: id of expense to return
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: GET expense by ID
      description: >-
        GET request using user input to select expense by ID from appropriate
        expenses table
      operationId: getExpenseById
      responses:
        '200':
          description: Default error sample response
          content:
            application/json:
              examples:
                single expense response:
                  value: >-
                    {id: 1, category: "Petrol", amount: 34, dateOfExpense:
                    10/10/23}
    delete:
      summary: DELETE expense by ID
      description: DELETE individual expense from user selected ID
      operationId: deleteById
      responses:
        '204':
          description: Default error sample response
          content:
            application/json:
              examples:
                expense deleted:
                  value: '{deleted: "Expense deleted successfully!"}'
    put:
      summary: UPDATE expense by ID
      description: Update an expense using user input, selected using expense unique ID
      operationId: updateExpense
      responses:
        '200':
          description: Default error sample response
          content:
            application/json:
              examples:
                expense updated:
                  value: '{updated: "Expense updated successfully"}'
...